WIP
WIP Initial conversion with all unit tests pass

WIP

WIP

Define more intefaces

Define the rest of interfaces

Clean up

WIP

WIP

Finish defining remaining interfaces

Clean up
